fix: added dashboard and history to containers
